% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{count_clusters}
\alias{count_clusters}
\title{Counts numbers of clusters in the defined period.}
\usage{
count_clusters(from_year, to_year)
}
\arguments{
\item{from_year}{years equal to or above.}

\item{to_year}{years lower or equal to.}
}
\value{
The number of clusters in the defined period.
}
\description{
Counts numbers of clusters in the defined period.
}
\examples{
count_clusters(1980, 2000)

}
