% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{graph_clusters}
\alias{graph_clusters}
\title{Graph clusters of municipality changes in Norway between \code{from_year} and \code{to_year}. Municipalities that remains unchanged, not merged or split, are not listed.}
\usage{
graph_clusters(from_year, to_year)
}
\arguments{
\item{from_year}{years equal to or above.}

\item{to_year}{years lower or equal to.}
}
\value{
A plot illustrating clusters for the defined period.
}
\description{
Graph clusters of municipality changes in Norway between \code{from_year} and \code{to_year}. Municipalities that remains unchanged, not merged or split, are not listed.
}
\examples{
graph_clusters(1980, 2000)

}
