% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link_municipality_names.R
\name{match_municipality}
\alias{match_municipality}
\title{Match Municipality Function}
\usage{
match_municipality(name, year = NULL, county_number = NULL, threshold = 0.1)
}
\arguments{
\item{name}{The name of the municipality.}

\item{year}{The year (optional). If not provided, the function will match the municipality
irrespective of the year.}

\item{county_number}{The number of the county (optional).}

\item{threshold}{The minimum difference in string distance between the best and the second best match.}
}
\value{
The number of the municipality that matches the provided name and (if specified) year and
county number, as well as the string distance and municipality name in a list.
}
\description{
This function takes a data.table with municipality details and matches a provided municipality name
to its corresponding municipality number for a given year, using Jaro-Winkler string matching method.
An optional county number can be specified.
}
\examples{
\dontrun{
df <- data.table(
  municipality_name = c("Municipality1", "Municipality2", "Municipality3"),
  municipality_number = c(1001, 1002, 1003),
  year_created = c(1900, 1920, 1940),
  year_stopped = c(1950, 1980, 2020)
)

name <- "Municipality1"
year <- 1930
county_num <- 10

match_municipality(df, name, year, county_num)
}
}
