% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NoRwayGeo-package.R
\docType{package}
\name{NoRwayGeo-package}
\alias{NoRwayGeo}
\alias{NoRwayGeo-package}
\title{NoRwayGeo: Makes Norwegian municipalities consistent over time}
\description{
Makes clusters of Norwegian municipalities that can be followed consistenyl over time.
}
\author{
\strong{Maintainer}: Eirik Berger \email{eirik.berger@gmail.com}

}
\keyword{internal}
