% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NoRwayGeo.R
\docType{package}
\name{NoRwayGeo}
\alias{NoRwayGeo}
\title{NoRwayGeo: A package for creating consistent municipality structures over time}
\description{
The NoRwayGeo package provides a packace for creating clusters of municipalities, such that its possible to consistently track areas over time.
}
