% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{print_changes}
\alias{print_changes}
\title{Print change in municipality structure in Norway between \code{from_year} and \code{to_year}.}
\usage{
print_changes(from_year, to_year)
}
\arguments{
\item{from_year}{years equal to or above.}

\item{to_year}{years lower or equal to.}
}
\value{
A data frame/data.table listing results.
}
\description{
Print change in municipality structure in Norway between \code{from_year} and \code{to_year}.
}
\examples{
print_changes(1980, 2000)

}
