% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{print_clusters}
\alias{print_clusters}
\title{Print clusters of municipalities in Norway between \code{from_year} and \code{to_year}. Municipalities that remains unchanged, not merged or split, are not listed.}
\usage{
print_clusters(from_year, to_year)
}
\arguments{
\item{from_year}{years equal to or above.}

\item{to_year}{years lower or equal to.}
}
\value{
A data frame/data.table listing municipality numbers with corresponding cluster ids.
}
\description{
Print clusters of municipalities in Norway between \code{from_year} and \code{to_year}. Municipalities that remains unchanged, not merged or split, are not listed.
}
\examples{
print_clusters(1980, 2000)

}
