% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NoRwayGeo-package.R
\docType{package}
\name{NoRwayGeo-package}
\alias{NoRwayGeo}
\alias{NoRwayGeo-package}
\title{NoRwayGeo: Dealing with Norwegian Municipalities}
\description{
Deals with Norwegian municipalities, including making them consistent across time and linking them to a municipality number.
}
\author{
\strong{Maintainer}: Eirik Berger \email{eirik.berger@gmail.com}

}
\keyword{internal}
